# docs/storage_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Use any http client to fetch data from the 0G Storage Scan.\n\n#### Common Error Messages:\nAn API call that encounters an error will return non-zero as its status code. The error message will be returned in the message field and the detailed reason for the error will be returned in the data field.\n```\n{\n\"code\": 1001,\n\"message\":\"No matching records found\",\n\"data\":\"Storage tx, txSeq 1000000\"\n}\n```\n| error code | error message             |\n|:-----------|:--------------------------|\n| 0          | Success                   |\n| 1          | Invalid parameter         |\n| 2          | Internal server error     |\n| 3          | Too many requests         |\n| 1001       | No matching records found |\n| 1002       | Blockchain RPC failure    |\n\n\n",
        "title": "0G Storage Scan API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/stats/storage": {
            "get": {
                "description": "Query data storage statistics, including incremental and full data, and support querying at hourly or daily time intervals",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistic"
                ],
                "summary": "Data storage statistics",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "maximum": 2000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp in seconds",
                        "name": "minTimestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp in seconds",
                        "name": "maxTimestamp",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "hour",
                            "day"
                        ],
                        "type": "string",
                        "default": "day",
                        "description": "Statistics interval",
                        "name": "intervalType",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort by timestamp",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.DataStatList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
...
        "/txs": {
            "get": {
                "description": "Query storage transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Storage transaction list",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The merkle root hash of the uploaded file",
                        "name": "rootHash",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The layer1 tx hash of the submission",
                        "name": "txHash",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.StorageTxList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BusinessError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "storage.DataStat": {
            "description": "Storage data information",
            "type": "object",
            "properties": {
                "dataSize": {
                    "description": "Size of storage data in a specific time interval",
                    "type": "integer"
                },
                "dataTotal": {
                    "description": "Total Size of storage data by a certain time",
                    "type": "integer"
                },
                "fileCount": {
                    "description": "Number of files in a specific time interval",
                    "type": "integer"
                },
                "fileTotal": {
                    "description": "Total number of files by a certain time",
                    "type": "integer"
                },
                "statTime": {
                    "description": "Statistics time",
                    "type": "string"
                }
            }
        },
        "storage.DataStatList": {
            "description": "Storage data list",
            "type": "object",
            "properties": {
                "list": {
                    "description": "Stat list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.DataStat"
                    }
                },
                "total": {
                    "description": "The total number of stat returned",
                    "type": "integer"
                }
            }
        },
...
        "storage.StorageTxInfo": {
            "description": "Submission transaction information",
            "type": "object",
            "properties": {
                "blockNumber": {
                    "description": "The block where the submit event is emitted",
                    "type": "integer"
                },
                "dataSize": {
                    "description": "File size in bytes",
                    "type": "integer"
                },
                "from": {
                    "description": "File uploader address",
                    "type": "string"
                },
                "method": {
                    "description": "The name of the submit event is always `submit`",
                    "type": "string"
                },
                "rootHash": {
                    "description": "Merkle root of the file to upload",
                    "type": "string"
                },
                "segments": {
                    "description": "The total number of segments the file is split into",
                    "type": "integer"
                },
                "status": {
                    "description": "File upload status, 0-not uploaded,1-uploading,2-uploaded",
                    "type": "integer"
                },
                "storageFee": {
                    "description": "The storage fee required to upload the file",
                    "type": "number"
                },
                "timestamp": {
                    "description": "The block time when submit event emits",
                    "type": "integer"
                },
                "txHash": {
                    "description": "The transaction where the submit event is emitted",
                    "type": "string"
                },
                "txSeq": {
                    "description": "Submission index in submit event",
                    "type": "integer"
                },
                "uploadedSegments": {
                    "description": "The number of segments the file has been uploaded",
                    "type": "integer"
                }
            }
        },
        "storage.StorageTxList": {
            "description": "Submission information list",
            "type": "object",
            "properties": {
                "list": {
                    "description": "Submission list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.StorageTxInfo"
                    }
                },
                "total": {
                    "description": "The total number of submission returned",
                    "type": "integer"
                }
            }
        },
...
    }
}